(PC-22884)[PRO] refactor: Refactored `refresh` usage.
