+++ pro/src/pages/AdageIframe/app/components/OffersInstantSearch/OffersSearch/Offers/NoResultsPage/NoResultsPage.tsx
@@ -2,8 +2,8 @@ import './NoResultsPage.scss'
 
 import React, { useContext } from 'react'
 
+import { ReactComponent as FullRefresh } from 'icons/full-refresh.svg'
 import { ReactComponent as SearchGrayIcon } from 'icons/ico-search-gray.svg'
-import { ReactComponent as ResetIcon } from 'icons/reset.svg'
 import {
   AlgoliaQueryContext,
   FacetFiltersContext,
@@ -28,7 +28,7 @@ export const NoResultsPage = ({
       {(query || facetFilters.length > 1) &&
         handleResetFiltersAndLaunchSearch && ( // la longueur de facetFilters doit être > 1 car il y aura toujours [offer.isEducational:true] dans les facetFilters
           <Button
-            Icon={ResetIcon}
+            Icon={FullRefresh}
             className="no-results-button"
             onClick={handleResetFiltersAndLaunchSearch}
             variant={ButtonVariant.SECONDARY}
