+++ pro/src/screens/Offers/SearchFilters/SearchFilters.tsx
@@ -16,7 +16,7 @@ import { Offerer, TSearchFilters } from 'core/Offers/types'
 import { hasSearchFilters } from 'core/Offers/utils'
 import { Audience } from 'core/shared'
 import { SelectOption } from 'custom_types/form'
-import { ReactComponent as ResetIcon } from 'icons/reset.svg'
+import { ReactComponent as FullRefresh } from 'icons/full-refresh.svg'
 import strokeCloseIcon from 'icons/stroke-close.svg'
 import { ButtonLink } from 'ui-kit'
 import { ButtonVariant } from 'ui-kit/Button/types'
@@ -215,7 +215,7 @@ const SearchFilters = ({
 
         <div className={styles['reset-filters']}>
           <ButtonLink
-            Icon={ResetIcon}
+            Icon={FullRefresh}
             {...resetFilterButtonProps}
             onClick={resetFilters}
             link={{
