+++ pro/src/ui-kit/Icons/Icons.tsx
@@ -33,6 +33,7 @@ const fullIcons: IconListItem[] = [
   { src: 'icons/full-hide.svg' }, // TODO still used in 1 ConfirmDialog
   { src: 'icons/full-back.svg' },
   { src: 'icons/full-download.svg' }, // TODO 3 uses left inside <Button> and <ButtonLink>
+  { src: 'icons/full-refresh.svg' }, // TODO still unsed in <Button>s
   { src: 'icons/full-parameters.svg' }, // TODO 1 use left inside a <Button>
   { src: 'icons/full-more.svg' }, // TODO still in 2 ButtonLink and 1 Button
   { src: 'icons/full-sort.svg' },
@@ -110,7 +111,6 @@ const notCleandIcons: IconListItem[] = [
   { src: 'icons/picto-tip.svg', viewBox: '0 0 18 18' },
   { src: 'icons/ico-radio-off.svg', viewBox: '0 0 16 16' },
   { src: 'icons/ico-radio-on.svg', viewBox: '0 0 16 16' },
-  { src: 'icons/reset.svg', viewBox: '0 0 20 20' },
   { src: 'icons/ico-right-arrow.svg', viewBox: '0 0 24 24' },
   { src: 'icons/ico-screen-play.svg', viewBox: '0 0 33 32' },
   { src: 'icons/ico-search-gray.svg', viewBox: '0 0 124 125' },
